Implement the main Reporter interface
This commit contains the main Reporter interface, base implementations,
and utility functions.  The base implementations are rootReporter, which
is used for the singletons returned by the Root and TODO utility
functions; and MockReporter, which allows for mocking out reporters.  The
utility function As is also implemented; modeled on errors.As, this
function allows retrieval of a specific reporter from a reporter chain.